backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Web3/Blockchain
web3>=6.0.0,<6.15.0
eth-account>=0.10.0,<0.11.0
eth-typing>=3.0.0,<4.0.0

# External Services
httpx==0.26.0
anthropic==0.18.1
# IPFS/Pinata - using httpx directly for API calls

# Caching
cachetools==5.3.2

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
testcontainers[postgres]>=4.0.0

# Utils
python-dotenv==1.0.1
